moved PizzaStores example into Factory Method solution
