+++ Factory-Method/Pizza-Stores/PizzaStore.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Factory_Pattern
+namespace Pizza_Stores
 {
     abstract class PizzaStore
     {
